requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
backoff==2.2.1
boto3==1.37.8
botocore==1.37.8
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
executing==2.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jiter==0.8.2
jmespath==1.0.1
jupyter-client==8.6.3
jupyter-core==5.7.2
levenshtein==0.27.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.65.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pre-commit
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-levenshtein==0.27.1
pytz==2025.1
pyzmq==26.2.1
rapidfuzz==3.12.2
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
